Added support for parsing XML with tree structures
This is basically a rewrite of sax_handler.rb to use a stack to keep
track of the position in the resulting datastructure while responding
to Nokogiri's SAX events. The former sax_hanlder.rb was not able to
handle parsing of tree structures like:

  <categories>
    <category id="1">
      <categories>
        <category id="2" />
      </categories>
    </category>
  </categories>

The new stack based aproach handles these cases - and as a bonus permits
passing a :class option to the 'element' method.